Unstale the repository
Repos are marked for auto-archival when the `Pushed_at` repository
attribute is older than a year. Repos can be permanently exempted from
the process by being added to an exclusion list.

This demo/documentation-oriented repository has been marked but is
important to keep open for a while. Permanent exclusion is not
desirable.

To prevent archival, this branch and commit are pushed. Marking should
happen again a year from now for us to reevaluate.

Should it need to be prolonged, this commit should be rewritten as-is
via `git commit --amend` then re-pushed to the `unstale` branch.